--- a/django/db/models/base.py
+++ b/django/db/models/base.py
@@ -407,91 +407,102 @@ class Model(metaclass=ModelBase):
         # overrides it. It should be one or the other; don't duplicate the work
         # The reason for the kwargs check is that standard iterator passes in by
         # args, and instantiation for iteration is 33% faster.
-        if len(args) > len(opts.concrete_fields):
-            # Daft, but matches old exception sans the err msg.
-            raise IndexError("Number of args exceeds number of fields")
+        concrete_fields_attnames_enumerated = opts.concrete_fields_attnames_enumerated
+        fields_len = len(concrete_fields_attnames_enumerated)
+        args_len = len(args)
 
-        if not kwargs:
-            fields_iter = iter(opts.concrete_fields)
-            # The ordering of the zip calls matter - zip throws StopIteration
-            # when an iter throws it. So if the first iter throws it, the second
-            # is *not* consumed. We rely on this, so don't change the order
-            # without changing the logic.
-            for val, field in zip(args, fields_iter):
+        if args_len == fields_len:
+            for i, attname in concrete_fields_attnames_enumerated:
+                val = args[i]
                 if val is _DEFERRED:
                     continue
-                _setattr(self, field.attname, val)
+                _setattr(self, attname, val)
         else:
-            # Slower, kwargs-ready version.
-            fields_iter = iter(opts.fields)
-            for val, field in zip(args, fields_iter):
-                if val is _DEFERRED:
+            if args_len > fields_len:
+                # Daft, but matches old exception sans the err msg.
+                raise IndexError("Number of args exceeds number of fields")
+
+            if not kwargs:
+                fields_iter = iter(opts.concrete_fields)
+                # The ordering of the zip calls matter - zip throws StopIteration
+                # when an iter throws it. So if the first iter throws it, the second
+                # is *not* consumed. We rely on this, so don't change the order
+                # without changing the logic.
+                for val, field in zip(args, fields_iter):
+                    if val is _DEFERRED:
+                        continue
+                    _setattr(self, field.attname, val)
+            else:
+                # Slower, kwargs-ready version.
+                fields_iter = iter(opts.fields)
+                for val, field in zip(args, fields_iter):
+                    if val is _DEFERRED:
+                        continue
+                    _setattr(self, field.attname, val)
+                    kwargs.pop(field.name, None)
+
+            # Now we're left with the unprocessed fields that *must* come from
+            # keywords, or default.
+
+            for field in fields_iter:
+                is_related_object = False
+                # Virtual field
+                if field.attname not in kwargs and field.column is None:
                     continue
-                _setattr(self, field.attname, val)
-                kwargs.pop(field.name, None)
-
-        # Now we're left with the unprocessed fields that *must* come from
-        # keywords, or default.
-
-        for field in fields_iter:
-            is_related_object = False
-            # Virtual field
-            if field.attname not in kwargs and field.column is None:
-                continue
-            if kwargs:
-                if isinstance(field.remote_field, ForeignObjectRel):
-                    try:
-                        # Assume object instance was passed in.
-                        rel_obj = kwargs.pop(field.name)
-                        is_related_object = True
-                    except KeyError:
+                if kwargs:
+                    if isinstance(field.remote_field, ForeignObjectRel):
                         try:
-                            # Object instance wasn't passed in -- must be an ID.
+                            # Assume object instance was passed in.
+                            rel_obj = kwargs.pop(field.name)
+                            is_related_object = True
+                        except KeyError:
+                            try:
+                                # Object instance wasn't passed in -- must be an ID.
+                                val = kwargs.pop(field.attname)
+                            except KeyError:
+                                val = field.get_default()
+                        else:
+                            # Object instance was passed in. Special case: You can
+                            # pass in "None" for related objects if it's allowed.
+                            if rel_obj is None and field.null:
+                                val = None
+                    else:
+                        try:
                             val = kwargs.pop(field.attname)
                         except KeyError:
+                            # This is done with an exception rather than the
+                            # default argument on pop because we don't want
+                            # get_default() to be evaluated, and then not used.
+                            # Refs #12057.
                             val = field.get_default()
-                    else:
-                        # Object instance was passed in. Special case: You can
-                        # pass in "None" for related objects if it's allowed.
-                        if rel_obj is None and field.null:
-                            val = None
                 else:
+                    val = field.get_default()
+
+                if is_related_object:
+                    # If we are passed a related instance, set it using the
+                    # field.name instead of field.attname (e.g. "user" instead of
+                    # "user_id") so that the object gets properly cached (and type
+                    # checked) by the RelatedObjectDescriptor.
+                    if rel_obj is not _DEFERRED:
+                        _setattr(self, field.name, rel_obj)
+                else:
+                    if val is not _DEFERRED:
+                        _setattr(self, field.attname, val)
+
+            if kwargs:
+                property_names = opts._property_names
+                for prop in tuple(kwargs):
                     try:
-                        val = kwargs.pop(field.attname)
-                    except KeyError:
-                        # This is done with an exception rather than the
-                        # default argument on pop because we don't want
-                        # get_default() to be evaluated, and then not used.
-                        # Refs #12057.
-                        val = field.get_default()
-            else:
-                val = field.get_default()
-
-            if is_related_object:
-                # If we are passed a related instance, set it using the
-                # field.name instead of field.attname (e.g. "user" instead of
-                # "user_id") so that the object gets properly cached (and type
-                # checked) by the RelatedObjectDescriptor.
-                if rel_obj is not _DEFERRED:
-                    _setattr(self, field.name, rel_obj)
-            else:
-                if val is not _DEFERRED:
-                    _setattr(self, field.attname, val)
-
-        if kwargs:
-            property_names = opts._property_names
-            for prop in tuple(kwargs):
-                try:
-                    # Any remaining kwargs must correspond to properties or
-                    # virtual fields.
-                    if prop in property_names or opts.get_field(prop):
-                        if kwargs[prop] is not _DEFERRED:
-                            _setattr(self, prop, kwargs[prop])
-                        del kwargs[prop]
-                except (AttributeError, FieldDoesNotExist):
-                    pass
-            for kwarg in kwargs:
-                raise TypeError("'%s' is an invalid keyword argument for this function" % kwarg)
+                        # Any remaining kwargs must correspond to properties or
+                        # virtual fields.
+                        if prop in property_names or opts.get_field(prop):
+                            if kwargs[prop] is not _DEFERRED:
+                                _setattr(self, prop, kwargs[prop])
+                            del kwargs[prop]
+                    except (AttributeError, FieldDoesNotExist):
+                        pass
+                for kwarg in kwargs:
+                    raise TypeError("'%s' is an invalid keyword argument for this function" % kwarg)
         super().__init__()
         post_init.send(sender=cls, instance=self)