@auvipy 2019-06-07
is this fall out of radar?